bridge/requirements.txt
bellows>=0.46.0
Pillow>=12.0.0
pycryptodome>=3.15.0
aiohttp>=3.9
websockets>=12.0
